pdfkit/requirements.txt
appdirs==1.4.3
avro==1.8.2
cycler==0.10.0
Flask==0.10.1
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.8
kafka-python==1.3.3
MarkupSafe==0.23
matplotlib==2.0.2
numpy==1.13.0rc1
packaging==16.8
pdfkit==0.6.1
pyparsing==2.2.0
pyPdf==1.13
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
subprocess32==3.2.7
Werkzeug==0.11.5
wkhtmltopdf==0.2
